Fix the formatting of certain C programs.

+++ languages/cprogs/Ex_1.5_reverse.c
@@ -22,7 +22,7 @@ int main(void)
 
     while(celsius >= lower)
     {
-        fahr = (9.0/5.0) * celsius + 32.0;
+        fahr = (float) ((9.0 / 5.0) * celsius + 32.0);
         printf("%3.0f %6.1f\n", celsius, fahr);
         celsius = celsius - step;
     }
